+++ src/theme/checkbox.ts
@@ -110,9 +110,10 @@ export default {
     },
     disabled: {
       true: {
-        base: 'cursor-not-allowed opacity-30',
-        label: 'cursor-not-allowed opacity-30',
-        description: 'cursor-not-allowed opacity-30'
+        root: 'opacity-30',
+        base: 'cursor-not-allowed',
+        label: 'cursor-not-allowed',
+        description: 'cursor-not-allowed'
       }
     },
     checked: {
@@ -133,8 +134,15 @@ export default {
       class: {
         root: 'border-(--b24ui-border-color) cursor-pointer'
       }
-    }
+    },
     // endregion ////
+    {
+      variant: 'card',
+      disabled: true,
+      class: {
+        root: 'cursor-not-allowed'
+      }
+    }
   ],
   defaultVariants: {
     color: 'air-primary',