+++ src/theme/radio-group.ts
@@ -145,8 +145,10 @@ export default {
     },
     disabled: {
       true: {
-        base: 'cursor-not-allowed opacity-30',
-        label: 'cursor-not-allowed opacity-30'
+        item: 'opacity-30',
+        base: 'cursor-not-allowed',
+        label: 'cursor-not-allowed',
+        description: 'cursor-not-allowed'
       }
     },
     required: {
@@ -222,8 +224,15 @@ export default {
         item: 'first-of-type:rounded-t-(--ui-border-radius-md) last-of-type:rounded-b-(--ui-border-radius-md)',
         fieldset: 'gap-0 -space-y-px'
       }
-    }
+    },
     // endregion ////
+    {
+      variant: ['card', 'table'],
+      disabled: true,
+      class: {
+        item: 'cursor-not-allowed'
+      }
+    }
   ],
   defaultVariants: {
     color: 'air-primary',